+++ libraries/PCA9671/CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.1] - 2025-03-18
+
+- fix #2, fix examples
+- remove address range check from getAddress()
+- add PCA9673 as derived class.
+- minor edits
+
+
 ## [0.1.0] - 2025-03-16
 
 - initial version based upon PCF8575 library.